requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
requests==2.26.0
python-dotenv==0.19.0
PyGithub==1.55
